third_party/blink/web_tests/platform/generic/external/wpt/scroll-animations/css/scroll-timeline-cssom.tentative-expected.txt

This is a testharness.js-based test.
Found 98 tests; 44 PASS, 54 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Empty block
PASS EOF ends block
PASS Timeline name can be a <string>
PASS Missing prelude
PASS Missing block
PASS Missing timeline name
PASS Timeline name must be an identifier
PASS Timeline name must match <custom-ident>
PASS Timeline name must match <custom-ident> (caps)
PASS Timeline name must match <custom-ident> (mixed)
PASS Timeline name may not be initial
PASS Timeline name may not be inherit
PASS Timeline name may not be unset
PASS Timeline name may not be revert
PASS Timeline name may not be default
PASS Extra timeline name
PASS CSSRule.type returns 0
PASS CSSScrollTimelineRule.name foo
PASS CSSScrollTimelineRule.name Foo
PASS CSSScrollTimelineRule.name f___123
PASS CSSScrollTimelineRule.name a\9 b
PASS CSSScrollTimelineRule.name "foo"
PASS CSSScrollTimelineRule.name "none"
PASS CSSScrollTimelineRule.cssText: empty rule
PASS CSSScrollTimelineRule.cssText: escaped name
PASS CSSScrollTimelineRule.cssText: source descriptor
PASS CSSScrollTimelineRule.cssText: orientation descriptor
FAIL CSSScrollTimelineRule.cssText: scroll-offsets descriptor (none) assert_equals: expected "@scroll-timeline timeline { scroll-offsets: none; }" but got "@scroll-timeline timeline { }"
FAIL CSSScrollTimelineRule.cssText: scroll-offsets descriptor (auto) assert_equals: expected "@scroll-timeline timeline { scroll-offsets: auto; }" but got "@scroll-timeline timeline { }"
FAIL CSSScrollTimelineRule.cssText: scroll-offsets descriptor (container-based offset, px) assert_equals: expected "@scroll-timeline timeline { scroll-offsets: 100px; }" but got "@scroll-timeline timeline { }"
FAIL CSSScrollTimelineRule.cssText: scroll-offsets descriptor (container-based offset, percentage) assert_equals: expected "@scroll-timeline timeline { scroll-offsets: 100%; }" but got "@scroll-timeline timeline { }"
FAIL CSSScrollTimelineRule.cssText: scroll-offsets descriptor (element offset) assert_equals: expected "@scroll-timeline timeline { scroll-offsets: selector(#bar); }" but got "@scroll-timeline timeline { }"
FAIL CSSScrollTimelineRule.cssText: scroll-offsets descriptor (element offset with edge) assert_equals: expected "@scroll-timeline timeline { scroll-offsets: selector(#bar) start; }" but got "@scroll-timeline timeline { }"
FAIL CSSScrollTimelineRule.cssText: scroll-offsets descriptor (element offset with threshold) assert_equals: expected "@scroll-timeline timeline { scroll-offsets: selector(#bar) 1; }" but got "@scroll-timeline timeline { }"
FAIL CSSScrollTimelineRule.cssText: scroll-offsets descriptor (element offset with edge and threshold) assert_equals: expected "@scroll-timeline timeline { scroll-offsets: selector(#bar) start 1; }" but got "@scroll-timeline timeline { }"
FAIL CSSScrollTimelineRule.cssText: scroll-offsets descriptor (element offset with threshold and edge) assert_equals: expected "@scroll-timeline timeline { scroll-offsets: selector(#bar) start 1; }" but got "@scroll-timeline timeline { }"
FAIL CSSScrollTimelineRule.cssText: scroll-offsets descriptor (multiple offsets) assert_equals: expected "@scroll-timeline timeline { scroll-offsets: auto, selector(#bar) start 1; }" but got "@scroll-timeline timeline { }"
FAIL CSSScrollTimelineRule.cssText: defaults assert_equals: expected "@scroll-timeline timeline { source: auto; orientation: auto; scroll-offsets: none; }" but got "@scroll-timeline timeline { source: auto; orientation: auto; }"
FAIL CSSScrollTimelineRule.cssText: order assert_equals: expected "@scroll-timeline timeline { source: auto; orientation: auto; scroll-offsets: none; }" but got "@scroll-timeline timeline { source: auto; orientation: auto; }"
PASS CSSScrollTimelineRule.source selector(#foo)
PASS CSSScrollTimelineRule.source selector( #foo )
PASS CSSScrollTimelineRule.source  selector(#foo) 
PASS CSSScrollTimelineRule.source none
PASS CSSScrollTimelineRule.source  none 
PASS CSSScrollTimelineRule.source selector(#a\9 b)
PASS CSSScrollTimelineRule.source auto
FAIL CSSScrollTimelineRule.source #foo assert_equals: serialization should be canonical expected "auto" but got "none"
FAIL CSSScrollTimelineRule.source  assert_equals: serialization should be canonical expected "auto" but got "none"
FAIL CSSScrollTimelineRule.source element(#foo) assert_equals: serialization should be canonical expected "auto" but got "none"
FAIL CSSScrollTimelineRule.source selector(#foo more) assert_equals: serialization should be canonical expected "auto" but got "none"
FAIL CSSScrollTimelineRule.source selector(html) assert_equals: serialization should be canonical expected "auto" but got "none"
FAIL CSSScrollTimelineRule.source selector(foo) assert_equals: serialization should be canonical expected "auto" but got "none"
FAIL CSSScrollTimelineRule.source selector(:before) assert_equals: serialization should be canonical expected "auto" but got "none"
FAIL CSSScrollTimelineRule.source selector(*) assert_equals: serialization should be canonical expected "auto" but got "none"
FAIL CSSScrollTimelineRule.source selector(.a) assert_equals: serialization should be canonical expected "auto" but got "none"
FAIL CSSScrollTimelineRule.source selector(.a, .b) assert_equals: serialization should be canonical expected "auto" but got "none"
PASS CSSScrollTimelineRule.orientation auto
PASS CSSScrollTimelineRule.orientation block
PASS CSSScrollTimelineRule.orientation inline
PASS CSSScrollTimelineRule.orientation horizontal
PASS CSSScrollTimelineRule.orientation vertical
PASS CSSScrollTimelineRule.orientation   vertical  
PASS CSSScrollTimelineRule.orientation 
PASS CSSScrollTimelineRule.orientation foo
PASS CSSScrollTimelineRule.orientation 10px
PASS CSSScrollTimelineRule.orientation red
FAIL CSSScrollTimelineRule.scrollOffsets none assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets  none  assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets  assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets red assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets #fff assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets unset assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets unknown(#foo) assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets start assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets end assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets 3 assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) 3 start 10 assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets 0%, 100%, selector(.a) assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) 3 end, start assert_equals: serialization should be canonical expected (string) "none" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets auto assert_equals: serialization should be canonical expected (string) "auto" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets 10em assert_equals: serialization should be canonical expected (string) "10em" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets 10% assert_equals: serialization should be canonical expected (string) "10%" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets calc(1px + 1%) assert_in_array: serialization should be sound value undefined not in array ["calc(1px + 1%)", "calc(1% + 1px)"]
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) assert_equals: serialization should be canonical expected (string) "selector(#foo)" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets  selector(#foo) assert_equals: serialization should be canonical expected (string) "selector(#foo)" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) start assert_equals: serialization should be canonical expected (string) "selector(#foo) start" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) start 3 assert_equals: serialization should be canonical expected (string) "selector(#foo) start 3" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) 3 assert_equals: serialization should be canonical expected (string) "selector(#foo) 3" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) 3.5 assert_equals: serialization should be canonical expected (string) "selector(#foo) 3.5" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) end assert_equals: serialization should be canonical expected (string) "selector(#foo) end" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) end 3 assert_equals: serialization should be canonical expected (string) "selector(#foo) end 3" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) 3 end assert_equals: serialization should be canonical expected (string) "selector(#foo) end 3" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets  auto , auto  assert_equals: serialization should be canonical expected (string) "auto, auto" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets 10%, 100px assert_equals: serialization should be canonical expected (string) "10%, 100px" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets auto, 100% assert_equals: serialization should be canonical expected (string) "auto, 100%" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets 0%, selector(  #foo)  3  end   assert_equals: serialization should be canonical expected (string) "0%, selector(#foo) end 3" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets selector(#foo) end 3, selector(#bar) assert_equals: serialization should be canonical expected (string) "selector(#foo) end 3, selector(#bar)" but got (undefined) undefined
FAIL CSSScrollTimelineRule.scrollOffsets 0%, auto, selector(#foo) start, auto, selector(#bar) 12.3 assert_equals: serialization should be canonical expected (string) "0%, auto, selector(#foo) start, auto, selector(#bar) 12.3" but got (undefined) undefined
Harness: the test ran to completion.
